feat(42): implement basic loading of show & hallplan infos

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/dto/hallplan/ShowHallplanSectionResponse.java
@@ -7,7 +7,7 @@ public record ShowHallplanSectionResponse(
     String name,
     String color,
     boolean standingOnly,
-    long price,
+    int price,
     String frontendCoordinates,
     int spotCount,
     int availableSpotCount,
